Add Storage, Ui, Parser and TaskList classes appropriately
